db: refactored repo interfaces
