--- Roki.Core/Services/Database/Repositories/IQuoteRepository.cs
@@ -1,14 +1,0 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using Roki.Core.Services.Database.Models;
-
-namespace Roki.Core.Services.Database.Repositories
-{
-    public interface IQuoteRepository : IRepository<Quote>
-    {
-        Task<Quote> GetRandomQuoteByKeywordAsync(ulong guildId, string keyword);
-        Task<Quote> SearchQuoteKeywordTextAsync(ulong guildId, string keyword, string text);
-        IEnumerable<Quote> GetGroup(ulong guildId, int page, OrderType order);
-        void RemoveAllByKeyword(ulong guildId, string keyword);
-    }
-}
